ArrumarDataHoraFormato

+++ cliente/deleteCliente.php
@@ -9,18 +9,28 @@ if (empty($Id)) {
 }
 
 $PDO = db_connect();
-$stmtCliente = $PDO->prepare($sqlCliente);
-$stmt->bindParam(':Id', $Id, PDO::PARAM_INT);
-$stmtCliente->execute();
-$sql = "DELETE FROM Cliente WHERE Id = :Id";
-$stmt = $PDO->prepare($sql);
-$stmt->bindParam(':Id', $Id, PDO::PARAM_INT);
 
-if ($stmt->execute()) {
-    header('Location: ../msg/msgSucesso.html');
-} else {
+// Verificar se existe alguma compra para este cliente
+$sqlCompra = "SELECT COUNT(*) AS total FROM Compra WHERE IdCliente = :Id";
+$stmtCompra = $PDO->prepare($sqlCompra);
+$stmtCompra->bindParam(':Id', $Id, PDO::PARAM_INT);
+$stmtCompra->execute();
+$total = $stmtCompra->fetchColumn();
+
+if ($total > 0) {
     header('Location: ../msg/msgErro.html');
-}
-exit;
+    exit;
 
+} else {
+    $sql = "DELETE FROM Cliente WHERE Id = :Id";
+    $stmt = $PDO->prepare($sql);
+    $stmt->bindParam(':Id', $Id, PDO::PARAM_INT);
+    
+    if ($stmt->execute()) {
+        header('Location: ../msg/msgSucesso.html');
+    } else {
+        header('Location: ../msg/msgErro.html');
+    }
+    exit;
+}
 ?>
